--- client/src/interfaces/IPost.tsx
@@ -1,9 +1,0 @@
-interface IPost {
-    post:{
-        title: string;
-        date: string;
-        status: string;
-        id: number
-    }
-}
-export default IPost;
