--- client/src/interfaces/IPostList.tsx
@@ -1,7 +1,0 @@
-import IPost from "./IPost";
-
-interface IPostList {
-    postList: IPost[];
-    name: string;
-}
-export default IPostList;
